Removed. Cleaned. Replaced unsafe strcpy by g_strlcpy.
	* glibtop_machine.h (skip_token): Removed.
	* glibtop_server.h (skip_multiple_token): Cleaned.
	* netload.c: (glibtop_get_netload_s): Replaced unsafe strcpy by g_strlcpy.
	* procmap.c: (glibtop_get_proc_map_s): s/strncpy/g_strlcpy/.
	* procstate.c: (glibtop_get_proc_state_s): Likewise.
	* sysinfo.c: (init_sysinfo): Cleaned.

+++ sysdeps/linux/glibtop_server.h
@@ -42,9 +42,9 @@ skip_token (const char *p)
 }
 
 static inline char *
-skip_multiple_token (const char *p, int count)
+skip_multiple_token (const char *p, size_t count)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < count; i++)
 		p = skip_token (p);
